LibI4/Python_AI/Extra/requirements.txt
voicevox-client>=*
speechrecognition>=*
asyncio>=*
pygame>=*
discord.py>=*
websockets>=*